Refactoring Type: Extract Function
https://refactoring.com/catalog/extractFunction.html

Put the inventory table in its own component and call it in the parent component Inventory
